+++ src/main/java/lotto/valid/InputValidator.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.valid;
 
 import java.util.ArrayList;
 import java.util.HashSet;
